+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 - 2.1.0
   - Added displaying of rule validation errors, see `showErrorMessage` in config.settings
+  - Added `QbUtils.isValidTree()`
   - `validateValue` moved from widget settings to field's `fieldSettings`
 - 2.0.11
   - Added `starts_with`, `ends_with` operators for text type